requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.15.0
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
trl>=0.7.1
wandb>=0.15.0
sentencepiece>=0.1.99
evaluate>=0.4.0
scikit-learn>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0.0
jupyter>=1.0.0
tqdm>=4.66.0
python-dotenv